Littlehope now filters against a template bank
Littlehope used to just filter against a template having the same masses
as the injection. Now, littlehope requires a template bank file as an
argument. It filters against every template described by the template
bank and then reports the template with the highest network SNR as a
coinc trigger.

To get the old behavior, use bayestar_sim_to_tmpltbank to create a
template bank from a lalapps_inspinj-style injection file.